+++ docs/changelog.rst
@@ -8,12 +8,15 @@ CHANGELOG
 **Bug fixes**
 
 - Fix: unable to search within a list of services (#3521)
-- Unpublish trek in all languages when path is deleted (#1321)
+- Fix: Unpublish trek in all languages when path is deleted (#1321)
+- Fix: To delete parent outdoor sites you must first delete their children (#3151)
+
 
 **New features**
 
 - Filter trek and outdoor site labels according to whether they are published or not (#3529)
 - Respond 404 JSON if page not found in API v2
+
 
 **Improvements**
 
@@ -29,6 +32,7 @@ CHANGELOG
 **Documentation**
 
 - Add configuration file for readthedocs
+
 
 
 2.99.0 (2023-07-18)